Update README prior to public release
Add ToC, Summary, References and what I hope is a clear
statement regarding potential future development of this
codebase.

At this point I feel like I've hit the high points for the
learning exercises and need to use (what little) time I
have to focus on next steps for the particular goals I
have in mind. Opening this repo up is not necessarily a
statement that my changes are particularly noteworthy, but
instead I'm hoping that the changes will be made visible to
anyone searching for help in completing a particular part of
the "Other Tasks" at the end of the article.

refs https://golang.org/doc/articles/wiki/